+++ fop-core/src/main/java/org/apache/fop/complexscripts/util/NumberConverter.java
@@ -254,7 +254,7 @@ public class NumberConverter {
         if (number < 0) {
             throw new IllegalArgumentException("number must be non-negative");
         } else if (token.length == 1) {
-            int s = token[0].intValue();
+            int s = token[0];
             switch (s) {
             case (int) '1':
                 fn = formatNumberAsDecimal(number, (int) '1', 1);
@@ -282,7 +282,7 @@ public class NumberConverter {
         } else if ((token.length == 2) && (token[0] == (int) 'W') && (token[1] == (int) 'w')) {
             fn = formatNumberAsWord(number, Character.TITLECASE_LETTER);
         } else if (isPaddedOne(token)) {
-            int s = token [ token.length - 1 ].intValue();
+            int s = token[token.length - 1];
             fn = formatNumberAsDecimal(number, s, token.length);
         } else {
             throw new IllegalArgumentException("invalid format token: \"" + UTF32.fromUTF32(token) + "\"");
@@ -503,7 +503,7 @@ public class NumberConverter {
     }
 
     private static int getDecimalValue(Integer scalar) {
-        int s = scalar.intValue();
+        int s = scalar;
         if (Character.getType(s) == Character.DECIMAL_DIGIT_NUMBER) {
             return Character.getNumericValue(s);
         } else {
